+++ @planetarium/tx/src/tx/index.ts
@@ -1,12 +1,6 @@
 export type { TxMetadata } from "./metadata.js";
 export { encodeTxMetadata } from "./metadata.js";
-export type {
-  UnsignedTxWithSystemAction,
-  UnsignedTxWithCustomActions,
-} from "./unsigned.js";
+export type { UnsignedTx } from "./unsigned.js";
 export type { SignedTx } from "./signed.js";
-export {
-  encodeUnsignedTxWithSystemAction,
-  encodeUnsignedTxWithCustomActions,
-} from "./unsigned.js";
+export { encodeUnsignedTx } from "./unsigned.js";
 export { signTx, encodeSignedTx } from "./signed.js";